feat: add LeadBottom component
