+++ packages/apps/crm/src/components/organisms/index.js
@@ -3,3 +3,4 @@ export {default as DropdownEventView} from './DropdownEventView/DropdownEventVie
 export {default as OpportunityDropdownInfo} from './OpportunityDropdownInfo/OpportunityDropdownInfo';
 export {default as LeadHeader} from './LeadHeader/LeadHeader';
 export {default as LeadBody} from './LeadBody/LeadBody';
+export {default as LeadBottom} from './LeadBottom/LeadBottom';
